feat: add cobra to the stack
